First attempts
I made a table for tic tac toe using html and css. I'm new to css, so I spent most of my time figuring that out. Now I'm moving on to java with changing the cells background image with the "onclick" script next. 
My confusion so far is that the code doesn't seem to work here, but works well in the w3schools.com "Try It" editor.